improve teh func to fuse init files and move/remove components
