+++ limbus_components_dev/base/__init__.py
@@ -1,2 +1,2 @@
-from .base import (ImageReader, Constant, Adder, ImageShow, Printer, Webcam, DrawBoxes,
-                   CropBoxes, Accumulator, ImageStitcher, ImageRegistrator)
+# the release script is not compatible with imports including ()
+from .base import ImageReader, Constant, Adder, ImageShow, Printer, Webcam, DrawBoxes, CropBoxes, Accumulator
