+++ limbus_components_dev/kornia/contrib/__init__.py
@@ -1,1 +1,2 @@
 from .contrib import ShowFaceLandmarks, FaceDetectorToBoxes
+from .contrib import ImageRegistrator, ImageStitcher
