+++ limbus_components_dev/kornia/contrib/contrib.py
@@ -143,3 +143,74 @@ class FaceDetectorToBoxes(Component):
         else:
             await self._outputs.boxes.send(torch.empty((0, 4, 2)).to(landmarks))
         return ComponentState.OK
+
+
+# temporal classes while we solve pending issues. TODO: allow components as parameters
+class ImageStitcher(Component):
+    """Component to stitch images together.
+
+    Args:
+        name (str): component name.
+        estimator (str): method to compute homography, either "vanilla" or "ransac".
+            Default: "ransac".
+        blending_method (str): method to blend two images together.
+            Only "naive" is currently supported. Default: "naive".
+
+    Inputs:
+        imgs (torch.Tensor): images to be stitched.
+
+    Outputs:
+        out (torch.Tensor): stitched image.
+
+    """
+    def __init__(self, name: str, estimator: str = 'ransac', blending_method: str = 'naive'):
+        super().__init__(name)
+        gftt_hardnet_matcher = kornia.feature.LocalFeatureMatcher(kornia.feature.GFTTAffNetHardNet(500),
+                                                                  kornia.feature.DescriptorMatcher('snn', 0.8))
+        self._is = kornia.contrib.ImageStitcher(gftt_hardnet_matcher, estimator, blending_method)
+
+    @staticmethod
+    def register_inputs(inputs: Params) -> None:  # noqa: D102
+        inputs.declare("imgs", torch.Tensor)
+
+    @staticmethod
+    def register_outputs(outputs: Params) -> None:  # noqa: D102
+        outputs.declare("out", torch.Tensor)
+
+    async def forward(self) -> ComponentState:  # noqa: D102
+        await self._outputs.out.send(self._is(*((await self._inputs.imgs.receive()).unsqueeze(1))))
+        return ComponentState.OK
+
+
+class ImageRegistrator(Component):
+    """Component to register images.
+
+    Args:
+        name (str): component name.
+
+    Inputs:
+        img_src (torch.Tensor): source image.
+        img_dst (torch.Tensor): destination image.
+
+    Outputs:
+        homo (torch.Tensor): homography.
+
+    """
+    def __init__(self, name: str):
+        super().__init__(name)
+        self._ir = kornia.geometry.ImageRegistrator()
+
+    @staticmethod
+    def register_inputs(inputs: Params) -> None:  # noqa: D102
+        inputs.declare("img_src", torch.Tensor)
+        inputs.declare("img_dst", torch.Tensor)
+
+    @staticmethod
+    def register_outputs(outputs: Params) -> None:  # noqa: D102
+        outputs.declare("homo", torch.Tensor)
+
+    async def forward(self) -> ComponentState:  # noqa: D102
+        img_src, img_dst = await asyncio.gather(self._inputs.img_src.receive(),
+                                                self._inputs.img_dst.receive())
+        await self._outputs.homo.send(self._ir.register(img_src, img_dst))
+        return ComponentState.OK
